pyproject.toml
[project]
name = "atomicads"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.4",
    "langgraph-checkpoint>=2.1.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
    "typer>=0.16.0",
    "uvicorn>=0.35.0",
]
